MNT: Rename schemacode -> bidsschematools
