+++ tools/schemacode/bidsschematools/__init__.py
@@ -12,6 +12,6 @@ __all__ = [
     "utils",
 ]
 
-version_file = files("schemacode.data") / "schema" / "SCHEMA_VERSION"
+version_file = files("bidsschematools.data") / "schema" / "SCHEMA_VERSION"
 with as_file(version_file) as path:
     __version__ = path.read_text().strip()
