+++ tools/schemacode/bidsschematools/data/schema
