+++ tools/schemacode/bidsschematools/render.py
