+++ tools/schemacode/bidsschematools/schema.py
