+++ tools/schemacode/bidsschematools/tests/data/broken_dataset_description.json
