+++ tools/schemacode/bidsschematools/tests/data/expected_bids_validator_xs_write.log
