+++ tools/schemacode/bidsschematools/tests/test_utils.py
@@ -1,4 +1,4 @@
-from schemacode import utils
+from bidsschematools import utils
 
 
 def test_combine_extensions():
