+++ tools/schemacode/bidsschematools/utils.py
