+++ tools/schemacode/pyproject.toml
@@ -1,5 +1,6 @@
 [build-system]
-requires = ["setuptools", "wheel"]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
 
 [tool.black]
 line-length = 99
@@ -17,8 +18,6 @@ exclude = '''
     | build
     | dist
   )/
-  | versioneer.py
-  | schemacode/_version.py
 )
 '''
