+++ tools/schemacode/setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
-name = schemacode
-version = file:schemacode/data/schema/SCHEMA_VERSION
-url = https://github.com/bids-standard/schemacode
+name = bidsschematools
+version = file:bidsschematools/data/schema/SCHEMA_VERSION
+url = https://github.com/bids-standard/bids-specification
 author = bids-standard developers
 author_email = bids.maintenance@gmail.com
 description = Python tools for working with the BIDS schema.
@@ -9,11 +9,10 @@ long_description = file:README.md
 long_description_content_type = text/markdown; charset=UTF-8; variant=GFM
 license = MIT
 classifiers =
-    Development Status :: 2 - Pre-Alpha
+    Development Status :: 3 - Alpha
     Intended Audience :: Science/Research
     Topic :: Scientific/Engineering :: Information Analysis
     License :: OSI Approved :: MIT License
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
@@ -48,17 +47,9 @@ all =
     %(tests)s
 
 [options.package_data]
-schemacode =
+bidsschematools =
     data/schema/SCHEMA_VERSION
     data/schema/**/*.yaml
-
-[versioneer]
-VCS = git
-style = pep440
-versionfile_source = schemacode/_version.py
-versionfile_build = schemacode/_version.py
-tag_prefix = v
-parentdir_prefix =
 
 [flake8]
 max-line-length = 99
